fine_tuning/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
duckdb==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==19.13.0
fastapi==0.112.4
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.43.0
gradio_client==1.3.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.4
InstructorEmbedding==1.0.1
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-experimental==0.3.0
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
llama-index-core==0.11.6
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-instructor==0.2.1
llama-index-llms-ollama==0.3.1
llama-index-readers-file==0.2.1
llama-index-readers-github==0.2.0
llama-index-storage-docstore-mongodb==0.2.0
llama-index-storage-index-store-mongodb==0.3.0
llama-index-storage-kvstore-mongodb==0.2.0
llama-index-vector-stores-mongodb==0.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.2.0
motor==3.5.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.6.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
ollama==0.3.2
openai==1.47.1
orjson==3.10.7
packaging==24.1
pandas==1.5.3
pandasai==2.2.14
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.3
plotly==5.24.1
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rich==13.8.0
rpds-py==0.20.0
ruff==0.6.4
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlglot==25.22.0
sqlglotrs==0.2.12
stack-data==0.6.3
starlette==0.38.4
streamlit==1.38.0
striprtf==0.0.26
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.9.11
zipp==3.20.1
